Made lib.browser.plugins.keepalive.handler doctests work in py3k
When the module was intitially ported, the doctest strings were not
updated to py3k. The documentation for the module should really be
moved over to ReST, but there's a good chance I'm going to drop the
module soon, as we'll likely be moving over to Requests, so it's
not worth the effort.